+++ src/adhocracy_core/adhocracy_core/rest/views.py
@@ -81,144 +81,6 @@ from adhocracy_core.workflows.schemas import create_workflow_meta_schema
 logger = getLogger(__name__)
 
 
-def validate_request_data(context: ILocation, request: Request,
-                          schema=MappingSchema(), extra_validators=[]):
-    """Validate request data.
-
-    :param context: passed to validator functions
-    :param request: passed to validator functions
-    :param schema: Schema to validate. Data to validate is extracted from the
-                   request.body. For schema nodes with attribute `location` ==
-                   `querystring` the data is extracted from the query string.
-                   The validated data (dict or list) is stored in the
-                   `request.validated` attribute.
-                   The `None` value is allowed to disable schema validation.
-    :raises HTTPBadRequest: HTTP 400 for bad request data.
-    """
-    body = {}
-    if request.content_type == 'multipart/form-data':
-        body = unflatten_multipart_request(request)
-    if request.content_type == 'application/json':
-        body = _extract_json_body(request)
-    validate_user_headers(request)
-    qs = _extract_querystring(request)
-    validate_body_or_querystring(body, qs, schema, context, request)
-    if request.errors:
-        request.validated = {}
-        raise HTTPBadRequest()
-
-
-def _extract_json_body(request: Request) -> object:
-    json_body = {}
-    if request.body == '':
-        request.body = '{}'
-    try:
-        json_body = request.json_body
-    except (ValueError, TypeError) as err:
-        error = error_entry('body', None,
-                            'Invalid JSON request body'.format(err))
-        request.errors.append(error)
-    return json_body
-
-
-def _extract_querystring(request: Request) -> dict:
-    parameters = {}
-    for key, value_encoded in request.GET.items():
-        import json
-        try:
-            value = json.loads(value_encoded)
-        except (ValueError, TypeError):
-            value = value_encoded
-        parameters[key] = value
-    return parameters
-
-
-def validate_user_headers(request: Request):
-    """
-    Validate the user headers.
-
-    If the request has a 'X-User-Path' and/or 'X-User-Token' header, we
-    ensure that the session takes belongs to the user and is not expired.
-    """
-    headers = request.headers
-    # TODO broken SRP, adhocracy_core.authentication is responsible instead
-    if 'X-User-Path' in headers or 'X-User-Token' in headers:
-        if get_user(request) is None:
-            error = error_entry('header', 'X-User-Token', 'Invalid user token')
-            request.errors.append(error)
-
-
-def validate_body_or_querystring(body, qs: dict, schema: MappingSchema,
-                                 context: IResource, request: Request):
-    """Validate the querystring if this is a GET request, the body otherwise.
-
-    This allows using just a single schema for all kinds of requests.
-    """
-    if isinstance(schema, GETPoolRequestSchema):
-        try:
-            schema = add_arbitrary_filter_nodes(qs,
-                                                schema,
-                                                context,
-                                                request.registry)
-        except Invalid as err:  # pragma: no cover
-            _add_colander_invalid_error_to_request(err, request,
-                                                   location='querystring')
-    if request.method.upper() == 'GET':
-        _validate_schema(qs, schema, request,
-                         location='querystring')
-    else:
-        _validate_schema(body, schema, request, location='body')
-
-
-def _validate_schema(cstruct: object, schema: MappingSchema, request: Request,
-                     location='body'):
-    """Validate that the :term:`cstruct` data is conform to the given schema.
-
-    :param request: request with list like `errors` attribute to append errors
-                    and the dictionary attribute `validated` to add validated
-                    data.
-    :param location: filter schema nodes depending on the `location` attribute.
-                     The default value is `body`.
-    """
-    if isinstance(schema, SequenceSchema):
-        _validate_list_schema(schema, cstruct, request, location)
-    elif isinstance(schema, MappingSchema):
-        _validate_dict_schema(schema, cstruct, request, location)
-    else:
-        error = 'Validation for schema {} is unsupported.'.format(str(schema))
-        raise(Exception(error))
-
-
-def _validate_list_schema(schema: SequenceSchema, cstruct: list,
-                          request: Request, location='body'):
-    if location != 'body':  # for now we only support location == body
-        return
-    child_cstructs = schema.cstruct_children(cstruct)
-    try:
-        request.validated = schema.deserialize(child_cstructs)
-    except Invalid as err:
-        _add_colander_invalid_error_to_request(err, request, location)
-
-
-def _validate_dict_schema(schema: MappingSchema, cstruct: dict,
-                          request: Request, location='body'):
-    validated = {}
-    try:
-        validated = schema.deserialize(cstruct)
-    except Invalid as err:
-        for child in err.children:
-            _add_colander_invalid_error_to_request(child, request, location)
-        if not err.children:
-            _add_colander_invalid_error_to_request(err, request, location)
-    request.validated.update(validated)
-
-
-def _add_colander_invalid_error_to_request(error: Invalid, request: Request,
-                                           location: str):
-    for name, msg in error.asdict().items():
-        request.errors.append(error_entry(location, name, msg))
-
-
 class RESTView:
     """Class stub with request data validation support.
